requirements.txt
pywebview==3.4
evdev==1.3.0
pyqtwebengine==5.15.5
pyqt5==5.15.6
